diffusers/requirements.txt
accelerate
torchvision
transformers>=4.25.1
datasets
ftfy
tensorboard
modelcards
bitsandbytes==0.35.0